Split main.py into main.py and epd_2in9_b.py
